Refactor WikipediaCollection class to have inner classes for Document
