--- src/main/java/io/anserini/document/WikipediaArticle.java
@@ -1,54 +1,0 @@
-/**
- * Anserini: An information retrieval toolkit built on Lucene
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.anserini.document;
-
-import java.io.BufferedReader;
-import java.io.IOException;
-
-/**
- * A Wikipedia article. The article title serves as the id.
- */
-public class WikipediaArticle implements SourceDocument {
-  private final String title;
-  private final String contents;
-
-  public WikipediaArticle(String title, String contents) {
-    this.title = title;
-    this.contents = contents;
-
-  }
-
-  @Override
-  public WikipediaArticle readNextRecord(BufferedReader bRdr) throws IOException {
-    return null;
-  }
-
-  @Override
-  public String id() {
-    return title;
-  }
-
-  @Override
-  public String content() {
-    return contents;
-  }
-
-  @Override
-  public boolean indexable() {
-    return true;
-  }
-}
