+++ src/test/java/io/anserini/document/WikipediaArticleTest.java
@@ -17,32 +17,26 @@
 package io.anserini.document;
 
 import java.io.BufferedOutputStream;
-import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.HashMap;
 
 import org.apache.commons.compress.compressors.bzip2.BZip2CompressorOutputStream;
 import org.apache.tools.ant.filters.StringInputStream;
-import org.apache.tools.bzip2.CBZip2InputStream;
-import org.apache.tools.bzip2.CBZip2OutputStream;
 import org.junit.Before;
 import org.junit.Test;
 
 import io.anserini.collection.Collection;
 import io.anserini.collection.WikipediaCollection;
 
-
-public class WikipediaArticleTest extends DocumentTest<WikipediaArticle> {
+public class WikipediaArticleTest extends DocumentTest<WikipediaCollection.Document> {
   protected static Path tmpPath;
 
   @Before
   public void setUp() throws Exception {
     super.setUp();
-    dType = new WikipediaArticle("Sample Title", "Sample Content");
+    dType = new WikipediaCollection.Document("Sample Title", "Sample Content");
 
     String doc = "<mediawiki xmlns=\"http://www.mediawiki.org/xml/export-0.10/\" xmlns:xsi=\"http://www.w3.org/2001/XMLSchema-instance\" xsi:schemaLocation=\"http://www.mediawiki.org/xml/export-0.10/ http://www.mediawiki.org/xm\n" +
         "l/export-0.10.xsd\" version=\"0.10\" xml:lang=\"en\">\n" +
